pyproject.toml
[project]
name = "surf-new"
version = "0.1.0"
description = "Surf new"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "annotated-types>=0.7.0",
    "anyio>=4.4.0",
    "certifi>=2024.7.4",
    "click>=8.1.7",
    "distro>=1.9.0",
    "dnspython>=2.6.1",
    "email_validator>=2.2.0",
    "fastapi>=0.111.1",
    "fastapi-cli>=0.0.4",
    "h11>=0.14.0",
    "httpcore>=1.0.5",
    "httptools>=0.6.1",
    "httpx>=0.27.0",
    "idna>=3.7",
    "Jinja2>=3.1.4",
    "langchain-core>=0.3.33",
    "langchain-anthropic>=0.1.23",
    "langchain-openai>=0.1.20",
    "markdown-it-py>=3.0.0",
    "MarkupSafe>=2.1.5",
    "mdurl>=0.1.2",
    "openai>=1.37.1",
    "pydantic>=2.10.2",
    "pydantic_core>=2.20.1",
    "playwright>=1.49.0",
    "Pygments>=2.18.0",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.9",
    "PyYAML>=6.0.1",
    "rich>=13.7.1",
    "shellingham>=1.5.4",
    "sniffio>=1.3.1",
    "starlette>=0.37.2",
    "tqdm>=4.66.4",
    "typer>=0.12.3",
    "typing_extensions>=4.12.2",
    "uvicorn>=0.30.3",
    "watchfiles>=0.22.0",
    "websockets>=12.0",
    "browser-use>=0.1.30",
    "langchain>=0.3.14",
    "psutil>=7.0.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.25.3",
    "pytest-cov>=6.0.0",
]
[tool.setuptools]
packages = ["api"]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
